requirements.txt
pyvesc == 1.0.5
pythoncrc == 1.21
pyserial-asyncio